+++ packages/@webex/plugin-meetings/test/unit/spec/meeting/index.js
@@ -69,6 +69,7 @@ import BEHAVIORAL_METRICS from '@webex/plugin-meetings/src/metrics/constants';
 import {MediaRequestManager} from '@webex/plugin-meetings/src/multistream/mediaRequestManager';
 import * as ReceiveSlotManagerModule from '@webex/plugin-meetings/src/multistream/receiveSlotManager';
 import * as SendSlotManagerModule from '@webex/plugin-meetings/src/multistream/sendSlotManager';
+import { CallDiagnosticUtils  } from '@webex/internal-plugin-metrics';
 
 import CallDiagnosticLatencies from '@webex/internal-plugin-metrics/src/call-diagnostic/call-diagnostic-metrics-latencies';
 import LLM from '@webex/internal-plugin-llm';
@@ -95,6 +96,16 @@ import {
   MeetingInfoV2PasswordError,
   MeetingInfoV2PolicyError,
 } from '../../../../src/meeting-info/meeting-info-v2';
+import {
+  CLIENT_ERROR_CODE_TO_ERROR_PAYLOAD,
+  DTLS_HANDSHAKE_FAILED_CLIENT_CODE,
+  ICE_FAILED_WITHOUT_TURN_TLS_CLIENT_CODE,
+  ICE_FAILED_WITH_TURN_TLS_CLIENT_CODE,
+  ICE_FAILURE_CLIENT_CODE,
+  MISSING_ROAP_ANSWER_CLIENT_CODE,
+} from '@webex/internal-plugin-metrics/src/call-diagnostic/config';
+import CallDiagnosticMetrics from '@webex/internal-plugin-metrics/src/call-diagnostic/call-diagnostic-metrics';
+import { ERROR_DESCRIPTIONS } from '@webex/internal-plugin-metrics/src/call-diagnostic/config';
 
 
 describe('plugin-meetings', () => {
@@ -1610,6 +1621,13 @@ describe('plugin-meetings', () => {
           const getErrorPayloadForClientErrorCodeStub = webex.internal.newMetrics.callDiagnosticMetrics.getErrorPayloadForClientErrorCode = sinon
             .stub()
             .returns(FAKE_ERROR);
+
+          const MOCK_CLIENT_ERROR_CODE = 2004;
+          const generateClientErrorCodeForIceFailureStub = sinon
+            .stub(CallDiagnosticUtils, 'generateClientErrorCodeForIceFailure')
+            .returns(MOCK_CLIENT_ERROR_CODE);
+
+
           meeting.meetingState = 'ACTIVE';
           meeting.mediaProperties.waitForMediaConnectionConnected.rejects(new Error('fake error'));
 
@@ -1623,7 +1641,12 @@ describe('plugin-meetings', () => {
               errorThrown = true;
             });
 
-          assert.calledOnceWithExactly(getErrorPayloadForClientErrorCodeStub, {clientErrorCode: 2004});
+          assert.calledOnceWithExactly(generateClientErrorCodeForIceFailureStub, {
+            signalingState: 'unknown',
+            iceConnectionState: 'unknown',
+            turnServerUsed: true,
+          })
+          assert.calledOnceWithExactly(getErrorPayloadForClientErrorCodeStub, {clientErrorCode: MOCK_CLIENT_ERROR_CODE});
           assert.calledTwice(webex.internal.newMetrics.submitClientEvent);
           assert.calledWithMatch(webex.internal.newMetrics.submitClientEvent, {
             name: 'client.ice.end',
@@ -1860,6 +1883,131 @@ describe('plugin-meetings', () => {
 
           await meeting.addMedia();
         });
+
+        describe('CA ice failures checks', () => {
+          [
+            {
+              clientErrorCode: ICE_FAILURE_CLIENT_CODE,
+              expectedErrorPayload: {
+                errorDescription: ERROR_DESCRIPTIONS.ICE_FAILURE,
+              },
+            },
+            {
+              clientErrorCode: MISSING_ROAP_ANSWER_CLIENT_CODE,
+              expectedErrorPayload: {
+                errorDescription: ERROR_DESCRIPTIONS.MISSING_ROAP_ANSWER,
+                category: 'signaling',
+              },
+            },
+            {
+              clientErrorCode: DTLS_HANDSHAKE_FAILED_CLIENT_CODE,
+              expectedErrorPayload: {
+                errorDescription: ERROR_DESCRIPTIONS.DTLS_HANDSHAKE_FAILED,
+              },
+            },
+            {
+              clientErrorCode: ICE_FAILED_WITHOUT_TURN_TLS_CLIENT_CODE,
+              expectedErrorPayload: {
+                errorDescription: ERROR_DESCRIPTIONS.ICE_FAILED_WITHOUT_TURN_TLS,
+              },
+            },
+            {
+              clientErrorCode: ICE_FAILED_WITH_TURN_TLS_CLIENT_CODE,
+              expectedErrorPayload: {
+                errorDescription: ERROR_DESCRIPTIONS.ICE_FAILED_WITH_TURN_TLS,
+              },
+            },
+          ].forEach(({clientErrorCode, expectedErrorPayload}) => {
+            it(`should handle all ice failures correctly for ${clientErrorCode}`, async () => {
+              // setting the method to the real implementation
+              // because newMetrics is mocked completely in the webex-mock
+              // the reason for this is that we want to test this on integration level
+              const CD = new CallDiagnosticMetrics({}, {parent: webex});
+              webex.internal.newMetrics.callDiagnosticMetrics.getErrorPayloadForClientErrorCode =
+                CD.getErrorPayloadForClientErrorCode;
+
+              const generateClientErrorCodeForIceFailureStub = sinon
+                .stub(CallDiagnosticUtils, 'generateClientErrorCodeForIceFailure')
+                .returns(clientErrorCode);
+
+              meeting.meetingState = 'ACTIVE';
+              meeting.mediaProperties.waitForMediaConnectionConnected.rejects(
+                new Error('fake error')
+              );
+
+              let errorThrown = false;
+
+              await meeting
+                .addMedia({
+                  mediaSettings: {},
+                })
+                .catch(() => {
+                  errorThrown = true;
+                });
+
+              assert.calledOnceWithExactly(generateClientErrorCodeForIceFailureStub, {
+                signalingState: 'unknown',
+                iceConnectionState: 'unknown',
+                turnServerUsed: true,
+              });
+
+              const submitClientEventCalls = webex.internal.newMetrics.submitClientEvent.getCalls();
+
+              assert.deepEqual(submitClientEventCalls[0].args, [
+                {
+                  name: 'client.media.capabilities',
+                  payload: {
+                    mediaCapabilities: {
+                      rx: {
+                        audio: false,
+                        share: false,
+                        share_audio: false,
+                        video: false,
+                        whiteboard: false,
+                      },
+                      tx: {
+                        audio: false,
+                        share: false,
+                        share_audio: false,
+                        video: false,
+                        whiteboard: false,
+                      },
+                    },
+                  },
+                  options: {
+                    meetingId: meeting.id,
+                  },
+                },
+              ]);
+
+              assert.deepEqual(submitClientEventCalls[1].args, [
+                {
+                  name: 'client.ice.end',
+                  payload: {
+                    canProceed: false,
+                    icePhase: 'JOIN_MEETING_FINAL',
+                    errors: [
+                      {
+                        fatal: true,
+                        shownToUser: false,
+                        name: 'other',
+                        category: 'media',
+                        errorCode: clientErrorCode,
+                        serviceErrorCode: undefined,
+                        ...expectedErrorPayload,
+                      },
+                    ],
+                  },
+                  options: {
+                    meetingId: meeting.id,
+                  },
+                },
+              ]);
+
+              assert.isTrue(errorThrown);
+            });
+          });
+        });
       });
 
       /* This set of tests are like semi-integration tests, they use real MuteState, Media, LocusMediaRequest and Roap classes.
@@ -2083,6 +2231,11 @@ describe('plugin-meetings', () => {
               iceServers,
             }, meetingId);
 
+            assert.calledWith(fakeMultistreamRoapMediaConnection.createSendSlot, MediaType.AudioMain, direction.audio !== 'inactive');
+            assert.calledWith(fakeMultistreamRoapMediaConnection.createSendSlot, MediaType.VideoMain, direction.video !== 'inactive');
+            assert.calledWith(fakeMultistreamRoapMediaConnection.createSendSlot, MediaType.AudioSlides, direction.screenShare !== 'inactive');
+            assert.calledWith(fakeMultistreamRoapMediaConnection.createSendSlot, MediaType.VideoSlides, direction.screenShare !== 'inactive');
+
             for(let type in localStreams){
               const stream = localStreams[type];
               if(stream !== undefined){
@@ -2111,7 +2264,7 @@ describe('plugin-meetings', () => {
                   screenShareVideo: localStreams.screenShareVideo?.outputTrack,
                   screenShareAudio: localStreams.screenShareAudio?.outputTrack,
                 },
-                direction,
+                direction: {audio: direction.audio, video: direction.video, screenShareVideo: direction.screenShare},
                 remoteQualityLevel,
               },
               expectedDebugId
@@ -2135,7 +2288,7 @@ describe('plugin-meetings', () => {
             direction: {
               audio: 'sendrecv',
               video: 'sendrecv',
-              screenShareVideo: 'recvonly',
+              screenShare: 'recvonly',
             },
             remoteQualityLevel: 'HIGH',
             expectedDebugId,
@@ -2165,7 +2318,7 @@ describe('plugin-meetings', () => {
             direction: {
               audio: 'sendrecv',
               video: 'sendrecv',
-              screenShareVideo: 'recvonly',
+              screenShare: 'recvonly',
             },
             remoteQualityLevel: 'HIGH',
             expectedDebugId,
@@ -2197,7 +2350,7 @@ describe('plugin-meetings', () => {
             direction: {
               audio: 'sendrecv',
               video: 'sendrecv',
-              screenShareVideo: 'recvonly',
+              screenShare: 'recvonly',
             },
             remoteQualityLevel: 'HIGH',
             expectedDebugId,
@@ -2226,7 +2379,7 @@ describe('plugin-meetings', () => {
             direction: {
               audio: 'inactive',
               video: 'sendrecv',
-              screenShareVideo: 'recvonly',
+              screenShare: 'recvonly',
             },
             remoteQualityLevel: 'HIGH',
             expectedDebugId,
@@ -2256,7 +2409,68 @@ describe('plugin-meetings', () => {
             direction: {
               audio: 'inactive',
               video: 'sendrecv',
-              screenShareVideo: 'recvonly',
+              screenShare: 'recvonly',
+            },
+            remoteQualityLevel: 'HIGH',
+            expectedDebugId,
+            meetingId: meeting.id
+          });
+
+          // and SDP offer was sent with the right audioMuted/videoMuted values
+          checkSdpOfferSent({audioMuted: true, videoMuted: true});
+
+          // and no other local mute requests were sent to Locus
+          assert.calledOnce(locusMediaRequestStub);
+        });
+
+
+        it('addMedia() works correctly when video is disabled with no tracks to publish', async () => {
+          await meeting.addMedia({videoEnabled: false});
+          await simulateRoapOffer();
+
+          // check RoapMediaConnection was created correctly
+          checkMediaConnectionCreated({
+            mediaConnectionConfig: expectedMediaConnectionConfig,
+            localStreams: {
+              audio: undefined,
+              video: undefined,
+              screenShareVideo: undefined,
+              screenShareAudio: undefined,
+            },
+            direction: {
+              audio: 'sendrecv',
+              video: 'inactive',
+              screenShare: 'recvonly',
+            },
+            remoteQualityLevel: 'HIGH',
+            expectedDebugId,
+            meetingId: meeting.id
+          });
+
+          // and SDP offer was sent with the right audioMuted/videoMuted values
+          checkSdpOfferSent({audioMuted: true, videoMuted: true});
+
+          // and no other local mute requests were sent to Locus
+          assert.calledOnce(locusMediaRequestStub);
+        });
+
+        it('addMedia() works correctly when screen share is disabled with no tracks to publish', async () => {
+          await meeting.addMedia({shareAudioEnabled: false, shareVideoEnabled: false});
+          await simulateRoapOffer();
+
+          // check RoapMediaConnection was created correctly
+          checkMediaConnectionCreated({
+            mediaConnectionConfig: expectedMediaConnectionConfig,
+            localStreams: {
+              audio: undefined,
+              video: undefined,
+              screenShareVideo: undefined,
+              screenShareAudio: undefined,
+            },
+            direction: {
+              audio: 'sendrecv',
+              video: 'sendrecv',
+              screenShare: 'inactive',
             },
             remoteQualityLevel: 'HIGH',
             expectedDebugId,